add events details page > heavy layout redo
